+++ packages/core/middleware/nuxt/index.js
@@ -8,16 +8,16 @@ module.exports = function VueStorefrontMiddleware(moduleOptions) {
   const options = {
     contentSecurityPolicy: false,
     crossOriginOpenerPolicy: false,
+    crossOriginEmbedderPolicy: false,
     permittedCrossDomainPolicies: {
       permittedPolicies: 'none'
     },
-    ...moduleOptions
+    ...moduleOptions.helmet || {}
   };
   // validate security setup with Helmet
   this.nuxt.hook('render:setupMiddleware', (app) => {
     app.use(cors());
-    const hasHelmetKey = Object.prototype.hasOwnProperty.call(options, 'enableHelmet');
-    const isHelmetEnabled = hasHelmetKey ? options.enableHelmet : true;
+    const isHelmetEnabled = moduleOptions.helmet || (moduleOptions.helmet && Object.keys(moduleOptions.helmet).length > 0) || false;
     if (isHelmetEnabled) {
       app.use(helmet(options));
       consola.success('Nuxt `Helmet` middleware added');